requirements.txt
aiohttp==3.8.4
aiosqlite==0.18.0
discord.py==2.2.2
python-dotenv==1.0.0
uvloop==0.17.0
waifuim.py==3.3.0
